requirements.txt
flet
google-generativeai